[REFACTOR(GuideLotto)] 수익률 매개변수 타입 변경

+++ src/main/kotlin/lotto/ui/GuideLotto.kt
@@ -19,7 +19,7 @@ class GuideLotto {
         )
     }
 
-    fun printIncomeRate(income: Float) {
+    fun printIncomeRate(income: Double) {
         println("총 수익률은 $income%입니다.")
     }
 }
